+++ lib/weeler/engine.rb
@@ -1,4 +1,5 @@
 module Weeler
+
   class Engine < ::Rails::Engine
     config.railties_order = [:main_app, Weeler::Engine, :all]
     config.weeler = Weeler
@@ -10,6 +11,7 @@ module Weeler
 
   ActiveSupport.on_load :action_controller do
     ActionDispatch::Routing::Mapper.send(:include, Weeler::RouteMapper)
+
   end
 
 end